+++ install/turbojet_control/share/turbojet_control/launch/turbojet_control.launch.py
@@ -7,7 +7,7 @@ def generate_launch_description():
     # Declare launch arguments
     port_name_arg = DeclareLaunchArgument(
         'port_name',
-        default_value='/dev/ttyUSB0',
+        default_value='/dev/ttyACM0',
         description='Serial port name'
     )